Update cachematrix.R
Created both functions, makeCacheMatrix() and cacheSolve().
The first one stores a matrix in cache and define four auxiliary functions to send and retrieve information from cache.

The second calculates the inverse of a matrix if it hasn't been calculated before. If it has, it just retrieves the results from cache.

More details in the file.